revert to float

+++ src/helper.cpp
@@ -83,7 +83,7 @@ void testResults(float gpu_result, Parameters &p) {
          "------------------------------\n");
 
   float EPS = 0.00001;
-  float x0 = (float)p.wos.x0.value;
+  float x0 = p.wos.x0.value;
 
   if (fabs(x0 - 0.0) < EPS) {
     float desired = /*0.042535;*/ 0.041561;
